moved modal functionality to admin dashboard instead of within the list / search, and removed dropdown menu in search
