+++ src/components/admin-dashboard/ItemDisplayList.jsx
@@ -3,7 +3,7 @@ import ItemModal from "../admin-dashboard/ItemModal";
 import convertDate from "../../utils/convertDate";
 import getAllItems from "../../utils/getAllItems";
 
-const ItemDisplayList = ({ item }) => {
+const ItemDisplayList = ({ item, onOpenModal }) => {
   const [showPreview, setShowPreview] = useState(false);
   const [modalItem, setModalItem] = useState(null);
   const [modalOpen, setModalOpen] = useState(false);
@@ -32,14 +32,8 @@ const ItemDisplayList = ({ item }) => {
 
   const handleItemClick = (item) => {
     setModalItem(item);
-    setModalOpen(true);
+    onOpenModal();
     console.log("Search query:", item.name);
-  };
-
-  const handleModalClose = () => {
-    setModalItem(null);
-    setModalOpen(false);
-    console.log("Modal closed.");
   };
 
   return (
@@ -58,38 +52,8 @@ const ItemDisplayList = ({ item }) => {
           Stock on hand: {item.stock}
         </p>
       </div>
-      {modalItem && modalOpen && (
-      <ItemModal item={modalItem} onClose={handleModalClose} />
-       )}
     </div>
   );
 };
 
-/*
-return (
-    <div className="flex border-[1px] justify-center align-center">
-      <div className="flex justify-center align-center">
-        <img src={item.imageUrl} className="max-w-[20vw] max-h-10vh" />
-      </div>
-      <div className="block border-2 border-red">
-        <p>
-          {item.name}, ${item.onSale ? item.price - item.discount : item.price}
-          <br />
-          original price: {item.price}, discount: {item.discount}
-        </p>
-        <small>
-          {item.onSale ? "On Sale" : "Not on Sale"},{" "}
-          {item.clearance ? "In Clearance Section" : "Not on Clearance."}
-        </small>
-        <br />
-        <small>Description: {item.description}</small>
-        <br />
-        <small>Added on: {date}</small>
-      </div>
-      <button>Edit this Item</button>
-      <button onClick={() => setShowPreview(true)}>Preview this Item</button>
-      <ItemDisplayModal open={showPreview} close={closePreview} item={item} />
-    </div>
-  );
-*/
 export default ItemDisplayList;
