+++ src/components/admin-dashboard/ItemSearch.jsx
@@ -3,7 +3,7 @@ import ItemModal from "./ItemModal";
 import AddItemModal from "./AddItemModal";
 import ItemDisplayList from "./ItemDisplayList";
 
-const ItemSearch = ({ items, updateItems }) => {
+const ItemSearch = ({ items, updateItems, onOpenModal }) => {
   const [search, setSearch] = useState("");
   const [searchResults, setSearchResults] = useState([]);
   const [noResults, setNoResults] = useState(false);
@@ -40,13 +40,12 @@ const ItemSearch = ({ items, updateItems }) => {
 
   const handleItemClick = (item) => {
     setModalItem(item);
-    setModalOpen(true);
+    onOpenModal();
     console.log("Search query:", item.name);
   };
 
   const handleModalClose = () => {
     setModalItem(null);
-    setSearch("");
     setModalOpen(false);
     setAddModalOpen(false);
     console.log("Modal closed.");
@@ -85,29 +84,8 @@ const ItemSearch = ({ items, updateItems }) => {
               {searchButtonText}
             </button>
           </form>
-          {search && (
-            <div className="text-left mt-2 absolute z-10 w-full bg-white border border-gray-300 shadow-md max-h-56 overflow-y-auto">
-              <ul className="py-1">
-                {searchResults.length > 0 ? (
-                  searchResults.slice(0, 10).map((item) => (
-                    <li
-                      key={item.id}
-                      onClick={() => handleItemClick(item)}
-                      className="px-4 py-2 hover:bg-gray-100 cursor-pointer"
-                    >
-                      {item.name}
-                    </li>
-                  ))
-                ) : (
-                  <li className="px-4 py-2">No results found.</li>
-                )}
-              </ul>
-            </div>
-          )}
+          
         </div>
-        {modalItem && modalOpen && (
-          <ItemModal item={modalItem} onClose={handleModalClose} />
-        )}
         {noResults && addModalOpen && (
           <AddItemModal item={search} onClose={handleModalClose} />
         )}
